+++ src/main/java/fuzs/goldenagecombat/client/handler/ClientCooldownHandler.java
@@ -13,24 +13,27 @@ public class ClientCooldownHandler {
     private AttackIndicatorStatus attackIndicator = AttackIndicatorStatus.OFF;
 
     @SubscribeEvent
-    public void onRenderGameOverlay(final RenderGameOverlayEvent evt) {
+    public void onRenderGameOverlay$Pre(final RenderGameOverlayEvent.Pre evt) {
         if (evt.getType() != RenderGameOverlayEvent.ElementType.ALL) return;
-        if (evt instanceof RenderGameOverlayEvent.Pre) {
-            if (GoldenAgeCombat.CONFIG.server().classic.removeCooldown) {
-                // this will mostly just remove the attack indicator, except for one niche case when looking at an entity
-                // just for that reason the whole indicator is also disabled later on
-                ClassicCombatHandler.disableCooldownPeriod(this.minecraft.player);
-            }
-            if (GoldenAgeCombat.CONFIG.client().classic.hideAttackIndicator) {
-                // indicator would otherwise render when looking at an entity, even when there is no cooldown
-                this.attackIndicator = this.minecraft.options.attackIndicator;
-                this.minecraft.options.attackIndicator = AttackIndicatorStatus.OFF;
-            }
-        } else if (evt instanceof RenderGameOverlayEvent.Post) {
-            if (GoldenAgeCombat.CONFIG.client().classic.hideAttackIndicator) {
-                // reset to old value
-                this.minecraft.options.attackIndicator = this.attackIndicator;
-            }
+        if (GoldenAgeCombat.CONFIG.server().classic.removeCooldown) {
+            // this will mostly just remove the attack indicator, except for one niche case when looking at an entity
+            // just for that reason the whole indicator is also disabled later on
+            ClassicCombatHandler.disableCooldownPeriod(this.minecraft.player);
+        }
+        if (GoldenAgeCombat.CONFIG.client().classic.hideAttackIndicator) {
+            // indicator would otherwise render when looking at an entity, even when there is no cooldown
+            this.attackIndicator = this.minecraft.options.attackIndicator;
+            this.minecraft.options.attackIndicator = AttackIndicatorStatus.OFF;
+        }
+    }
+
+    @SubscribeEvent
+    public void onRenderGameOverlay$Post(final RenderGameOverlayEvent.Post evt) {
+        if (evt.getType() != RenderGameOverlayEvent.ElementType.ALL) return;
+        if (GoldenAgeCombat.CONFIG.client().classic.hideAttackIndicator) {
+            // reset to old value
+            // we don't just leave this disabled as it'll change the vanilla setting permanently, which a mod shouldn't do imo
+            this.minecraft.options.attackIndicator = this.attackIndicator;
         }
     }