+++ src/main/java/fuzs/goldenagecombat/mixin/client/HumanoidModelMixin.java
@@ -25,15 +25,15 @@ public abstract class HumanoidModelMixin<T extends LivingEntity> extends Ageable
     public void setupAnim(T entityIn, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch, CallbackInfo callbackInfo) {
         if (entityIn instanceof Player player) {
             if (SwordBlockingHandler.isActiveItemStackBlocking(player)) {
-                final boolean oldBlockingPose = GoldenAgeCombat.CONFIG.client().animations.oldBlockingPose;
+                final boolean simpleBlockingPose = GoldenAgeCombat.CONFIG.client().animations.simpleBlockingPose;
                 if (entityIn.getUsedItemHand() == InteractionHand.OFF_HAND) {
                     this.leftArm.xRot = this.leftArm.xRot - ((float) Math.PI * 2.0F) / 10.0F;
-                    if (!oldBlockingPose) {
+                    if (simpleBlockingPose) {
                         this.leftArm.yRot = ((float) Math.PI / 6.0F);
                     }
                 } else {
                     this.rightArm.xRot = this.rightArm.xRot - ((float) Math.PI * 2.0F) / 10.0F;
-                    if (!oldBlockingPose) {
+                    if (simpleBlockingPose) {
                         this.rightArm.yRot = ((float) -Math.PI / 6.0F);
                     }
                 }